Fix crash on pane click.
cmd_mouse_at was fine without adjusting for the status location. The function cares
mostly about the location of the cursor within the actual pane and currently is only
used for copy mode tracking. As such, allowing it to enter the area of the pane status
causes problems without adding any functionality.